Reorganized and cleaned
Reorganized the control and setup classes, cleaned unnecessary redundancy
