--- Calculator/src/Conversion/ConversionSetup.java
@@ -1,14 +1,0 @@
-package Conversion;
-
-import java.util.Scanner;
-
-import Control.EasterEggs;
-
-public class ConversionSetup {	
-	public Scanner scanner = new Scanner(System.in);
-	public int forgive = 1;
-	public double input;
-	public double rslt;
-	public int order;
-	public EasterEggs ee = new EasterEggs();
-}
