--- Calculator/src/Conversion/Moles/MoleSetup.java
@@ -1,16 +1,0 @@
-package Conversion.Moles;
-import java.util.Scanner;
-
-import Control.EasterEggs;
-
-public class MoleSetup {
-    public Scanner scanner = new Scanner(System.in);
-    public int action;
-    public double molarmass;
-    public double moles;
-    public double molecs;
-    public double mass;
-    public String cont;
-    public int forgive = 1;
-    public EasterEggs ee = new EasterEggs();
-}
