+++ Calculator/src/Math/Addition.java
@@ -1,5 +1,7 @@
 package Math;
 
+import Setup.CalculatorSetup;
+
 public class Addition extends CalculatorSetup {
 	public void run () {
 			System.out.println("Input:\n");
