--- Calculator/src/Math/CalculatorSetup.java
@@ -1,16 +1,0 @@
-package Math;
-
-import java.util.Scanner;
-
-import Control.EasterEggs;
-
-public class CalculatorSetup {
-	public EasterEggs ee = new EasterEggs();
-	public double x;
-	public double y;
-	public double z;
-	public int forgive = 1;
-	public String cont = "Y";
-	public double rslt;
-	public Scanner scanner = new Scanner(System.in);
-}
