+++ Calculator/src/Setup/CalculatorSetup.java
@@ -1,0 +1,14 @@
+package Setup;
+
+import java.util.Scanner;
+
+public class CalculatorSetup {
+	public EasterEggs ee = new EasterEggs();
+	public double x;
+	public double y;
+	public double z;
+	public int forgive = 1;
+	public String cont = "Y";
+	public double rslt;
+	public Scanner scanner = new Scanner(System.in);
+}
