+++ Calculator/src/Setup/ConversionSetup.java
@@ -1,0 +1,12 @@
+package Setup;
+
+import java.util.Scanner;
+
+public class ConversionSetup {	
+	public Scanner scanner = new Scanner(System.in);
+	public int forgive = 1;
+	public double input;
+	public double rslt;
+	public int order;
+	public EasterEggs ee = new EasterEggs();
+}
