+++ Calculator/src/Setup/EasterEggs.java
@@ -1,0 +1,13 @@
+package Setup;
+
+public class EasterEggs {
+    public void cknfngr () {
+        System.out.println("\n\n\nEAT MORE CHICKEN\n\n\n");
+    }
+    public void woof () {
+        System.out.println("\n\n\nWOOF WOOF!\n\n\n");
+    }
+    public void dirty () {
+        System.out.println("\n\n\nGet your head out of the gutter\n\n\n");
+    }
+}
