+++ Calculator/src/Setup/MoleSetup.java
@@ -1,0 +1,15 @@
+package Setup;
+
+import java.util.Scanner;
+
+public class MoleSetup {
+    public Scanner scanner = new Scanner(System.in);
+    public int action;
+    public double molarmass;
+    public double moles;
+    public double molecs;
+    public double mass;
+    public String cont;
+    public int forgive = 1;
+    public EasterEggs ee = new EasterEggs();
+}
